cancel and update campaigns

+++ app/Http/Controllers/Staff/MohCampaignController.php
@@ -9,16 +9,17 @@ use App\Models\Campaign;
 use App\Models\Hospital;
 use App\Models\User;
 use Symfony\Contracts\Service\Attribute\Required;
+use \Carbon\Carbon;
 
 class MohCampaignController extends Controller {
     //# Get all campaigns
-    public function manageCampaigns(Request $request) {
+    public function index(Request $request) {
         $campaigns = Campaign::get();
         return view('moh.manage-campaigns')->with('campaigns', $campaigns);
     }
 
     //# Add new campaign
-    public function addCampaign(Request $request) {
+    public function create(Request $request) {
         if ($request->end_date < $request->start_date)
             return back()->with('message', 'End date cannot be before start date');
         if (!$request->location)
@@ -30,14 +31,14 @@ class MohCampaignController extends Controller {
             foreach ($request->doctors as $doctor) {
                 $user = User::where('national_id', $doctor)->first();
                 if (!$user)
-                    return redirect('/staff/moh/manage-campaigns')->with('message', 'Doctor with ID ' . $doctor . ' does not exist');
+                    return back()->with('message', 'Doctor with ID ' . $doctor . ' does not exist');
 
                 //# Check if doctor is already working in a campaign
                 if ($user->campaigns()->first()) {
                     $busy_doctor = $user->campaigns()->where('start_date', '>', now())->first();
                     $unavailable_doctor = $user->campaigns()->where('end_date', '>', now())->first();
                     if ($busy_doctor || $unavailable_doctor)
-                        return redirect('/staff/moh/manage-campaigns')->with('message', 'Doctor with ID ' . $doctor . ' is already assigned to a campaign');
+                        return back()->with('message', 'Doctor with ID ' . $doctor . ' is already assigned to a campaign');
                 }
             }
         }
@@ -62,17 +63,55 @@ class MohCampaignController extends Controller {
         }
 
         if ($campaign)
-            return redirect('/staff/moh/manage-campaigns')->with('message', 'Campaign added successfully');
+            return back()->with('message', 'Campaign added successfully');
         else
-            return redirect('/staff/moh/manage-campaigns')->with('message', 'Campaign could not be added');
+            return back()->with('message', 'Campaign could not be added');
     }
 
     public function cancel(Campaign $campaign) {
+        if (now()->diffInDays($campaign->start_date) < 1) {
+            return back()->with('message', 'Can\'t cancel a campaign that is starting in two days !');
+        }
+
         $campaign->appointments()->update(['campaign_appointments.status' => 'cancelled']);
         $campaign->delete();
 
         //TODO: we should probably notify the user that their campaign has been cancelled
 
-        return redirect('/staff/moh/manage-campaigns')->with('message', 'Campaign cancelled successfully');
+        return back()->with('message', 'Campaign cancelled successfully');
+    }
+    
+    public function update(Request $request, Campaign $campaign) {
+        if (now()->diffInDays($campaign->start_date) < 1) {
+            return back()->with('message', 'Can\'t update a campaign that is starting in two days !');
+        }
+
+        $request->validate([
+            'start_date' => 'required',
+            'end_date' => 'required',
+            'location' => 'required',
+            'city' => 'required',
+            'address' => 'required',
+            'capacity_per_day' => ['required', 'numeric', 'min:1'],
+        ]);
+
+        $campaign->update([
+            'start_date' => $request->start_date,
+            'end_date' => $request->end_date,
+            'location' => $request->location,
+            'city' => $request->city,
+            'address' => $request->address,
+            'capacity_per_day' => $request->capacity_per_day,
+        ]);
+
+        return redirect('/staff/moh/manage-campaigns')->with('message', 'Campaign updated successfully');
+    }
+
+    public function updateView(Request $request, Campaign $campaign) {
+        if (now()->diffInDays($campaign->start_date) < 1) {
+            return back()->with('message', 'Can\'t update a campaign that is starting in two days !');
+        }
+
+        return view('moh.update-campaign')->with('campaign', $campaign);
     }
 }
